\name{mixPHM-package}
\alias{mixPHM-package}
\alias{mixPHM}
\docType{package}
\title{
Mixtures of proportional hazard models
}
\description{
This package fits multiple variable mixtures of various parametric proportional hazard models using the EM-Algorithm. 
Proportionality restrictions can be imposed on the latent groups and/or on the variables. 
Several survival distributions can be specified. Missing and censored values are allowed. Independence is assumed over the single variables.
}
\details{
\tabular{ll}{
Package: \tab mixPHM\cr
Type: \tab Package\cr
Version: \tab 0.7-2\cr
Date: \tab 2015-07-23\cr
License: \tab GPL-2\cr
}
}
\author{
Patrick Mair, Marcus Hudec

Maintainer: Patrick Mair <mair@fas.harvard.edu>
}
\references{
Mair, P., and Hudec, M. (2009). Multivariate Weibull mixtures with proportional hazard
restrictions for dwell time based session clustering with incomplete data. Journal of
the Royal Statistical Society, Series C (Applied Statistics), 58(5), 619-639.

Kalbfleisch, J.D., and Prentice, R.L. (1980). The statistical analysis of failure 
time data. New York: Wiley. 

Celaux, G., and Govaert, G. (1992). A classification EM algorithm for clustering and
two stochastic versions. Computational Statistics and Data Analysis, 14, 315-332.
}

\keyword{ package }

