\name{mixPHM-package}
\alias{mixPHM-package}
\alias{mixPHM}
\docType{package}
\title{
Mixtures of proportional hazard models
}
\description{
This package fits multiple variable mixtures of various parametric proportional hazard models using the EM-Algorithm. 
Proportionality restrictions can be imposed on the latent groups and/or on the variables. 
Several survival distributions can be specified. Missing values are allowed. Independence is assumed over the single variables.
}
\details{
\tabular{ll}{
Package: \tab mixPHM\cr
Type: \tab Package\cr
Version: \tab 0.6.2\cr
Date: \tab 2008-06-26\cr
License: \tab GPL\cr
}
}
\author{
Patrick Mair, Marcus Hudec

Maintainer: Patrick Mair <patrick.mair@wu-wien.ac.at>
}
\references{
Mair, P., & Hudec, M. (2008). Analysis of dwell times in Web Usage Mining. In C. 
Preisach, H. Burkhardt, L. Schmidt-Thieme, & R. Decker (eds.), Data Analysis, Machine 
Learning, and Applications, pp. 593-600. Berlin: Springer.

Kalbfleisch, J.D., and Prentice, R.L. (1980). The statistical analysis of failure 
time data. New York: Wiley. 

Celaux, G., and Govaert, G. (1992). A classification EM algorithm for clustering and
two stochastic versions. Computational Statistics and Data Analysis, 14, 315-332.
}

\keyword{ package }

